fix: ensure end event is triggered before deploy
